move main activity
