+++ app/src/main/AndroidManifest.xml
@@ -16,6 +16,11 @@
         <activity
             android:name=".MovieActivity"
             android:exported="false" />
+
+        <activity
+            android:name=".HomeActivity"
+            android:exported="false" />
+
         <activity
             android:name=".MainActivity"
             android:theme="@style/Theme.Mufiest.NoActionBar"
@@ -26,6 +31,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
     </application>
 
 </manifest>
