+++ app/src/main/res/drawable/login_image_gradient_overlay.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android" >
+    <!-- Top gradient -->
+    <item android:bottom="1400px">
+        <shape android:shape="rectangle">
+            <gradient
+                android:startColor="#0D0D0D"
+                android:endColor="#00000000"
+                android:angle="270"/>
+        </shape>
+    </item>
+
+    <!-- Left gradient -->
+    <item android:right="500px">
+        <shape android:shape="rectangle">
+            <gradient
+                android:startColor="#0D0D0D"
+                android:endColor="#00000000"
+                android:angle="0"/>
+        </shape>
+    </item>
+
+    <!-- Right gradient -->
+    <item android:left="500px">
+        <shape android:shape="rectangle">
+            <gradient
+                android:startColor="#0D0D0D"
+                android:endColor="#00000000"
+                android:angle="180"/>
+        </shape>
+    </item>
+
+    <!-- Bottom gradient -->
+    <item android:top="1400px">
+        <shape android:shape="rectangle">
+            <gradient
+                android:startColor="#0D0D0D"
+                android:endColor="#00000000"
+                android:angle="90"/>
+        </shape>
+    </item>
+</layer-list>
