+++ app/src/main/res/layout/activity_main.xml
@@ -1,38 +1,86 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
+    android:layout_height="match_parent"
+    android:background="#0D0D0D"
     tools:context=".MainActivity">
 
-    <LinearLayout
+    <!-- Background Image -->
+    <ImageView
         android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_centerInParent="true"
+        android:scaleType="centerCrop"
+        android:contentDescription="Login Background Image"
+        android:src="@drawable/image_login_page" />
+
+    <!-- Gradient Overlay -->
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@drawable/login_image_gradient_overlay" />
+
+    <!-- Content -->
+    <RelativeLayout
+        android:id="@+id/centerContainer"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:orientation="vertical" >
+        android:layout_centerInParent="true">
 
-        <LinearLayout
-            android:layout_width="match_parent"
+        <!-- Mufiest Logo -->
+        <ImageView
+            android:id="@+id/logoImageView"
+            android:layout_width="250dp"
+            android:layout_height="56dp"
+            android:src="@drawable/mufiest_upper_logo_gold"
+            android:contentDescription="Mufiest Logo"
+            android:layout_centerHorizontal="true"/>
+
+        <EditText
+            android:id="@+id/editTextEmail"
+            android:layout_width="330dp"
             android:layout_height="wrap_content"
-            android:id="@+id/popularMovieListContainer"
-            android:layout_marginTop="20dp"
-            android:orientation="horizontal" />
+            android:layout_below="@id/logoImageView"
+            android:layout_marginTop="16dp"
+            android:layout_marginBottom="16dp"
+            android:hint="Email"
+            android:inputType="textEmailAddress"
+            android:background="#FFFFFF"/>
 
-        <LinearLayout
-            android:layout_width="match_parent"
+        <EditText
+            android:id="@+id/editTextPassword"
+            android:layout_width="330dp"
             android:layout_height="wrap_content"
-            android:id="@+id/favoriteMovieListContainer"
-            android:layout_marginTop="5dp"
-            android:orientation="horizontal" />
+            android:layout_below="@id/editTextEmail"
+            android:layout_marginBottom="16dp"
+            android:hint="Password"
+            android:inputType="textPassword"
+            android:background="#FFFFFF"/>
 
+        <!-- Login Button -->
         <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/recentReviewListContainer"
-            android:layout_marginTop="5dp"
-            android:orientation="horizontal" />
+            android:id="@+id/loginButton"
+            android:layout_width="100dp"
+            android:layout_height="43dp"
+            android:orientation="horizontal"
+            android:gravity="center"
+            android:background="@drawable/rounded_button"
 
-    </LinearLayout>
+            android:layout_below="@id/editTextPassword"
+            android:layout_centerHorizontal="true"
+            android:layout_margin="2dp">
 
-</ScrollView>
+            <!-- Login Button -->
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textColor="#000000"
+                android:text="Login"
+                android:textSize="16dp" />
+        </LinearLayout>
+
+    </RelativeLayout>
+
+</RelativeLayout>
